requirements.txt
numpy
pandas
sklearn
xgboost
tensorflow-gpu
keras
matplotlib
seaborn
flask
